add OccupancyTreeMapDisplay

+++ include/erl_geometry/octree_impl.tpp
@@ -422,7 +422,8 @@ namespace erl::geometry {
         const uint32_t keyval = std::floor(coordinate * m_resolution_inv_);
         const uint32_t diff = tree_depth - depth;
         if (!diff) { return keyval + m_tree_key_offset_; }
-        return ((keyval >> diff) << diff) + (1 << (diff - 1)) + m_tree_key_offset_;
+        return ((keyval >> diff) << diff) + static_cast<uint32_t>(1 << (diff - 1)) +
+               m_tree_key_offset_;
     }
 
     template<class Node, class Interface, class InterfaceSetting>
@@ -1018,7 +1019,7 @@ namespace erl::geometry {
         if (this->m_stack_.empty()) { return; }
         ERL_DEBUG_ASSERT(tree != nullptr, "Tree is null.");
 
-        // If the provided AABB is too large, CoordsToKeyChecked will return false. We should avoid
+        // If the provided AABB is too large, CoordToKeyChecked will return false. We should avoid
         // this case.
         if (!this->GetInTreeAabb(
                 aabb_min_x,
